refactor(erc20-collateral-pools): add new version v2
