+++ src/pools/index.ts
@@ -1,4 +1,5 @@
 export * from './admin-pools'
 export * from './v1.erc20-collateral-pool'
 export * from './v1.erc20-collateral-pool-toplend'
+export * from './v2.erc20-collateral-pool'
 export * from './pools'
